Feature: Add custom region shapes (#28)
This pull request introduces support for adding shapefiles of non-standard regions in a standardized manner.

Standard regions continue to be sourced from `cartopy.io.shapereader.natural_earth`.

The shapefiles of all non-standard regions are placed inside a folder called `shapes` along with the respective scripts to generate them.

Additionally, this pull request adds a script to plot a global map of countries and regions for which data is available.